# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yemen Flow Hub API",
        "description": "Internal regulatory workflow API for CBY import financing requests.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current API server"
        }
    ],
    "paths": {
...
        "/api/requests/{importRequest}/documents": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload request document",
                "operationId": "f387a6eae03daaed34a7b89d02031520",
                "parameters": [
                    {
                        "name": "importRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document uploaded"
                    }
                }
            }
        },
        "/api/workflow/{importRequest}/swift-upload": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload SWIFT document and trigger workflow transition",
                "operationId": "93f6e04e3f1097215a75ce37a720f25f",
                "parameters": [
                    {
                        "name": "importRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "SWIFT uploaded"
                    }
                }
            }
        },
...
        "/api/requests": {
            "get": {
                "tags": [
                    "Import Requests"
                ],
                "summary": "List requests",
                "operationId": "81f070558de3bee7a1acf5f707a02794",
                "responses": {
                    "200": {
                        "description": "Requests retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Import Requests"
                ],
                "summary": "Create draft request",
                "operationId": "f4e7970c60149062c38cb7a8104c22d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merchant_id",
                                    "currency",
                                    "amount",
                                    "supplier_name",
                                    "goods_description",
                                    "port_of_entry"
                                ],
                                "properties": {
                                    "merchant_id": {
                                        "type": "integer"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "enum": [
                                            "USD",
                                            "EUR",
                                            "SAR",
                                            "AED",
                                            "CNY"
                                        ]
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "supplier_name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "goods_description": {
                                        "type": "string"
                                    },
                                    "port_of_entry": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            }
        },
        "/api/requests/{id}": {
            "get": {
                "tags": [
                    "Import Requests"
                ],
                "summary": "Show request details",
                "operationId": "932a4b4d28771c0e9526071a67c261f7",
                "responses": {
                    "200": {
                        "description": "Request retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Import Requests"
                ],
                "summary": "Update request",
                "operationId": "91a21fac2e0dc46c6322f7ae74553371",
                "responses": {
                    "200": {
                        "description": "Request updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Import Requests"
                ],
                "summary": "Delete draft request",
                "operationId": "b837f7524126dbc3717f488790c5321c",
                "responses": {
                    "200": {
                        "description": "Request deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            }
        },
        "/api/requests/{id}/history": {
            "get": {
                "tags": [
                    "Workflow"
                ],
                "summary": "Request stage history",
                "operationId": "2381e42554faf01229090aa1c47a9eb5",
                "responses": {
                    "200": {
                        "description": "History retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "sanctumCookie": []
                    }
                ]
            }
        },
...
        "/api/notifications/{notification}/read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark one notification as read",
                "operationId": "0335843b4658cbd12e3f19136665e4dd",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked read"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctumCookie": {
                "type": "apiKey",
                "description": "Sanctum stateful cookie auth for SPA",
                "name": "laravel_session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum personal access token",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}